fix manager
