+++ gsyVideoPlayer-java/src/main/java/com/shuyu/gsyvideoplayer/GSYVideoManager.java
@@ -2,9 +2,21 @@ package com.shuyu.gsyvideoplayer;
 
 
 import android.annotation.SuppressLint;
+import android.content.Context;
+import android.view.View;
+import android.view.ViewGroup;
+import android.view.Window;
+
+import com.danikula.videocache.HttpProxyCacheServer;
 import com.shuyu.gsyvideoplayer.listener.GSYMediaPlayerListener;
+import com.shuyu.gsyvideoplayer.utils.CommonUtil;
+
+import java.io.File;
 
 import tv.danmaku.ijk.media.player.IjkLibLoader;
+
+import static com.shuyu.gsyvideoplayer.utils.CommonUtil.hideNavKey;
+import static com.shuyu.gsyvideoplayer.video.base.GSYBaseVideoPlayer.FULLSCREEN_ID;
 
 
 /**
@@ -64,6 +76,78 @@ public class GSYVideoManager extends GSYVideoBaseManager {
      */
     public static synchronized void changeManager(GSYVideoManager gsyVideoManager) {
         videoManager = gsyVideoManager;
+    }
+
+
+    /**
+     * 获取缓存代理服务
+     */
+    protected static HttpProxyCacheServer getProxy(Context context) {
+        HttpProxyCacheServer proxy = GSYVideoManager.instance().proxy;
+        return proxy == null ? (GSYVideoManager.instance().proxy =
+                GSYVideoManager.instance().newProxy(context)) : proxy;
+    }
+
+
+    /**
+     * 获取缓存代理服务,带文件目录的
+     */
+    public static HttpProxyCacheServer getProxy(Context context, File file) {
+
+        //如果为空，返回默认的
+        if (file == null) {
+            return getProxy(context);
+        }
+
+        //如果已经有缓存文件路径，那么判断缓存文件路径是否一致
+        if (GSYVideoManager.instance().cacheFile != null
+                && !GSYVideoManager.instance().cacheFile.getAbsolutePath().equals(file.getAbsolutePath())) {
+            //不一致先关了旧的
+            HttpProxyCacheServer proxy = GSYVideoManager.instance().proxy;
+
+            if (proxy != null) {
+                proxy.shutdown();
+            }
+            //开启新的
+            return (GSYVideoManager.instance().proxy =
+                    GSYVideoManager.instance().newProxy(context, file));
+        } else {
+            //还没有缓存文件的或者一致的，返回原来
+            HttpProxyCacheServer proxy = GSYVideoManager.instance().proxy;
+
+            return proxy == null ? (GSYVideoManager.instance().proxy =
+                    GSYVideoManager.instance().newProxy(context, file)) : proxy;
+        }
+    }
+
+    /**
+     * 退出全屏，主要用于返回键
+     *
+     * @return 返回是否全屏
+     */
+    @SuppressWarnings("ResourceType")
+    public static boolean backFromWindowFull(Context context) {
+        boolean backFrom = false;
+        ViewGroup vp = (ViewGroup) (CommonUtil.scanForActivity(context)).findViewById(Window.ID_ANDROID_CONTENT);
+        View oldF = vp.findViewById(FULLSCREEN_ID);
+        if (oldF != null) {
+            backFrom = true;
+            hideNavKey(context);
+            if (GSYVideoManager.instance().lastListener() != null) {
+                GSYVideoManager.instance().lastListener().onBackFullscreen();
+            }
+        }
+        return backFrom;
+    }
+
+    /**
+     * 页面销毁了记得调用是否所有的video
+     */
+    public static void releaseAllVideos() {
+        if (GSYVideoManager.instance().listener() != null) {
+            GSYVideoManager.instance().listener().onCompletion();
+        }
+        GSYVideoManager.instance().releaseMediaPlayer();
     }
